Completed README.md commented R file to generate codebook fragment Optimized data set loading

+++ run_analysis.R
@@ -1,7 +1,9 @@
-#
-# this are variables for tidy data
-#
-#
+###
+# Coursera, Getting And Cleaning Daza,
+# Programming Assignment, July 2014
+###
+
+# load libraries
 library(plyr)
 library(reshape2)
 
@@ -12,6 +14,7 @@ dataDir <- paste(mainDir, subDir, sep = "/")
 pathToTxtFile <- "features.txt"
 pathToZipFile <- "getdata-projectfiles-UCI HAR Dataset.zip"
 fileUrl <- "https://d396qusza40orc.cloudfront.net/getdata%2Fprojectfiles%2FUCI%20HAR%20Dataset.zip"
+doDownloadFile <- TRUE
 
 ###
 # try to download the ZIPed data file from the given url
@@ -30,25 +33,40 @@ if (!file.exists(dataDir)) {
 # this file ist just one representer we could use a better
 # test condition here, but it should work....
 if(!file.exists(paste(dataDir, pathToTxtFile, sep = "/"))) {
-  # no, so download the ZIP from url specified above
+  # no, so download the ZIP if configured from url specified above
   zipFilePath <- paste(dataDir, pathToZipFile, sep = "/")
-  download.file(fileUrl, destfile=zipFilePath)
-  # and unzip the data file
-  unzip(zipFilePath, exdir=dataDir, overwrite=TRUE)
-  # remove the ZIP, it's no longer needed
-  file.remove(zipFilePath)
+  if(doDownloadFile == TRUE) {
+    # only download the data set from the internet if configured
+    # otherwise the downlaoded file will be awaited in the data folder
+    # configured by the 'subDir' varriable
+    download.file(fileUrl, destfile=zipFilePath)
+    
+    message("Data file downloaded successfully.")
+  }
   
-  # save the target dir where the unziped files will be stored at least
-  # for later use
-  unzipedTargetDir = paste(dataDir, "UCI_HAR_Dataset", sep="/")
-  # remove spaces from the directory where the data files were unziped
-  file.rename(paste(dataDir, "UCI HAR Dataset", sep="/"), unzipedTargetDir)
-  # copy all unziped files one folder back (../)
-  dirContent <- list.files(unzipedTargetDir)
-  fileList <- paste(unzipedTargetDir, dirContent, sep="/")
-  file.copy(fileList, dataDir, overwrite=TRUE, recursive=TRUE)
-  # remove the unzipedTargetDir folder and all it's sub directories
-  unlink(unzipedTargetDir, recursive=TRUE, force=TRUE)
+  if(!file.exists(paste(dataDir, pathToZipFile, sep = "/"))) {
+    stop("Script execution aborted, reason: 'getdata-projectfiles-UCI HAR Dataset.zip' not found in data dir.")
+  }
+  else {
+    # and unzip the data file
+    unzip(zipFilePath, exdir=dataDir, overwrite=TRUE)
+    # remove the ZIP, it's no longer needed
+    file.remove(zipFilePath)
+  
+    # save the target dir where the unziped files will be stored at least
+    # for later use
+    unzipedTargetDir = paste(dataDir, "UCI_HAR_Dataset", sep="/")
+    # remove spaces from the directory where the data files were unziped
+    file.rename(paste(dataDir, "UCI HAR Dataset", sep="/"), unzipedTargetDir)
+    # copy all unziped files one folder back (../)
+    dirContent <- list.files(unzipedTargetDir)
+    fileList <- paste(unzipedTargetDir, dirContent, sep="/")
+    file.copy(fileList, dataDir, overwrite=TRUE, recursive=TRUE)
+    # remove the unzipedTargetDir folder and all it's sub directories
+    unlink(unzipedTargetDir, recursive=TRUE, force=TRUE)
+    
+    message("Data files unziped and prepared successfully.")
+  }
 }
 
 # load list of all data files in the data dir creared above
@@ -57,6 +75,8 @@ allFilesdf <- data.frame(files=list.files(dataDir, recursive = TRUE), stringsAsF
 removeFiles <- grepl("Inertial", allFilesdf$files)
 # get a list of relevant files
 selectedFiles <- allFilesdf[!removeFiles, ]
+
+message("Loading data sets, please be patient...")
 
 ###
 # Step 1: read all necessary data
@@ -81,6 +101,7 @@ featuresHead <- read.table(paste(dataDir, selectedFiles[2], sep="/"))
 # load "data/activity_labels.txt"
 activityLabels <- read.table(paste(dataDir, selectedFiles[1], sep="/"), col.names=c("activityId", "activity"), stringsAsFactors=FALSE)
 
+message("Raw data sets loaded...")
 
 ###
 # Step 2: Combining
@@ -95,6 +116,8 @@ combinedPerformers <- rbind(trainPerformerId, testPerformerId)
 
 # combine train and test data without performers (subjects) and activities
 combinedDataSet <- rbind(trainX, testX)
+
+message("Raw data sets combined...")
 
 ###
 # Step 3: Cleaning dataset
@@ -124,6 +147,8 @@ filteredHeadNamesCleaned <- gsub("BodyBody", "Body", filteredHeadNamesCleaned)
 #filteredHeadNamesCleaned <- gsub("(-x|-y|-z)", "\\U\\1", filteredHeadNamesCleaned, perl=TRUE)
 filteredHeadNamesCleaned <- gsub("-", "", filteredHeadNamesCleaned)
 
+
+message("Data set cleaned up...")
 
 ###
 # Step 4: building the final set
@@ -158,6 +183,8 @@ completeDataSet <- arrange(completeDataSet, performerId, activity)
 # rearrange column: moving performerId- and activity-columns to the beginning of the data.frame
 completeDataSet <- cbind(completeDataSet[, 67:68], completeDataSet[, -(67:68)])
 
+message("Final cleaned up data set created...")
+
 
 #######
 ## Step 6: create tidy dataset
@@ -167,6 +194,8 @@ tidyDataSet <- ddply(completeDataSet, .(activity, performerId), colwise(mean))
 # sorting
 tidyDataSet <- arrange(tidyDataSet, performerId, activity)
 
+message("Tidy data set created...")
+
 #####
 # Step 7: write the complete and tidy data set out to disk
 #####
@@ -174,3 +203,5 @@ tidyDataSet <- arrange(tidyDataSet, performerId, activity)
 write.table(completeDataSet, "uci_har_dataset_merged.txt")
 # write out the tidy data set to disk
 write.table(tidyDataSet, "uci_har_tidy_dataset.txt")
+
+message("Written data sets to disk and all done.")
